+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 - Fix instance_count when using draw_index with instance buffers
 - Added a `reinterpret` function to `BufferSlice`
 - Made `AttributeInfo` derive `Copy`, `Clone` and `Debug`
-
 - Use [google/shaderc](https://github.com/google/shaderc-rs) for shader compilation
+- Reject generation of rust types for SPIR-V arrays that would have incorrect array stride.
 
 # Version 0.10.0 (2018-08-10)